+++ src/pages/channel/ui/channel-list.tsx
@@ -1,9 +1,9 @@
-import CreateButton from './create-button';
+import ChannelAddButton from './channel-add-button';
 
 const ChannelList = () => {
   return (
     <div className='grid grid-cols-2 gap-3'>
-      <CreateButton />
+      <ChannelAddButton />
     </div>
   );
 };
